docs/examples/foobar_api/requirements.txt
flask-marshmallow-openapi
